Flatten hierarchy.
